Move generators to their own module.
